# [Flag Database] Fix performance issue