--- a/src/main/java/org/openstreetmap/atlas/checks/database/FlagDatabaseSubCommand.java
+++ b/src/main/java/org/openstreetmap/atlas/checks/database/FlagDatabaseSubCommand.java
@@ -377,7 +377,6 @@ public class FlagDatabaseSubCommand extends AbstractAtlasShellToolsCommand
 
                         counter += features.size();
                     }
-
                 }
             }
             // Execute the remaining features